requirements.txt
nextcord>=2.0.0
python-telegram-bot>=13.0
python-dotenv>=0.19.0
